fix(01.1): resolve ruff lint findings across the library
- RUF013 (130): annotate default-None parameters as X | None across 23
  modules; full diff reviewed, every hunk is a pure '| None' insertion
- B006 (3): projected_gradient dict defaults moved to None + in-body
  materialisation; call-compatible for every existing caller
- B026 (4): star-args moved ahead of keywords in forwarding calls --
  binding-identical (positional binding always precedes keyword binding);
  sample alias verified bit-identical to rvs
- F841 (10) dead locals removed; RUF059 (16) unused unpacks renamed to
  underscore form; B007 (2) loop variables anonymised; B905 zip gains
  strict=True behind an existing equal-length guard
- RUF046 (2) redundant int(math.ceil(...)); RUF005 tuple splat; RUF015
  next(iter(...)); UP007 Scalar alias to PEP 604 (isinstance-equivalence
  proven for all member types)
- E731 (11): fit-internal lambdas converted to named defs (EM scan
  steps, generator closures, cross-stage Hessian helper)
- E501 (41): docstrings/comments wrapped; LaTeX reflowed at whitespace
  (render-identical); four module docstrings drop filler prefix
- per-file-ignores with rationale: E741 waived where l IS the GH-family
  lambda parameter (McNeil et al notation); RUF022 waived for the
  family-grouped timeseries __all__
- library paths now ruff-clean under the D-14 rule set; 756 module tests
  pass across every owning suite; test-file findings remain for the
  test-suite lint plan

+++ copulax/_src/timeseries/_unit_root.py
@@ -148,7 +148,7 @@ def _schwert_lags(n: int) -> int:
     so the result is a Python ``int`` consumed as a static-loop bound
     inside the compiled graph.
     """
-    return int(math.ceil(12.0 * (n / 100.0) ** 0.25))
+    return math.ceil(12.0 * (n / 100.0) ** 0.25)
 
 
 def _interp_p_jit(
@@ -404,7 +404,7 @@ def kpss(
         # difference is at most one bandwidth and the Bartlett kernel
         # with adjacent ``l`` values gives test statistics that differ
         # by only a few parts in 10^4.
-        lags = int(math.ceil(coef * (n / 100.0) ** 0.25))
+        lags = math.ceil(coef * (n / 100.0) ** 0.25)
     lags = int(lags)
 
     cols: list[Array] = [jnp.ones((n,), dtype=float)]
